requirements.txt

Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0
google-genai==0.8.0
requests==2.31.0
gTTS==2.5.4
moviepy==2.2.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
pexels-api-py==1.0.0
Pillow==10.0.1
scipy==1.11.4
numpy==1.24.4
gunicorn==21.2.0
python-dateutil==2.8.2
uuid==1.30
pathlib2==2.3.7
black==23.7.0
flake8==6.0.0
pytest==7.4.2
pytest-flask==1.3.0
cryptography==41.0.7
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4